meteor: Set up a "generate-globals" script to automatically generate global typings from module typings.
Fixes #15071 (as best we can without a massive breaking change).
