+++ types/meteor/globals/README.md
@@ -1,0 +1,3 @@
+This `globals/` directory is generated by `scripts/generate-globals`!  To update
+this directory, edit the corresponding files outside `globals/` and re-run
+`scripts/generate-globals`.
