+++ types/meteor/globals/reactive-var.d.ts
@@ -1,0 +1,8 @@
+declare var ReactiveVar: ReactiveVarStatic;
+declare interface ReactiveVarStatic {
+    new <T>(initialValue: T, equalsFunc?: Function): ReactiveVar<T>;
+}
+declare interface ReactiveVar<T> {
+    get(): T;
+    set(newValue: T): void;
+}
