+++ types/meteor/index.d.ts
@@ -12,24 +12,42 @@
 // TypeScript Version: 2.8
 
 /// <reference path="./accounts-base.d.ts" />
+/// <reference path="./globals/accounts-base.d.ts" />
 /// <reference path="./blaze.d.ts" />
+/// <reference path="./globals/blaze.d.ts" />
 /// <reference path="./browser-policy-common.d.ts" />
+/// <reference path="./globals/browser-policy-common.d.ts" />
 /// <reference path="./check.d.ts" />
+/// <reference path="./globals/check.d.ts" />
 /// <reference path="./ddp-rate-limiter.d.ts" />
+/// <reference path="./globals/ddp-rate-limiter.d.ts" />
 /// <reference path="./ddp.d.ts" />
+/// <reference path="./globals/ddp.d.ts" />
 /// <reference path="./ejson.d.ts" />
+/// <reference path="./globals/ejson.d.ts" />
 /// <reference path="./email.d.ts" />
+/// <reference path="./globals/email.d.ts" />
 /// <reference path="./http.d.ts" />
+/// <reference path="./globals/http.d.ts" />
 /// <reference path="./meteor.d.ts" />
+/// <reference path="./globals/meteor.d.ts" />
 /// <reference path="./mongo.d.ts" />
+/// <reference path="./globals/mongo.d.ts" />
 /// <reference path="./random.d.ts" />
+/// <reference path="./globals/random.d.ts" />
 /// <reference path="./reactive-var.d.ts" />
+/// <reference path="./globals/reactive-var.d.ts" />
 /// <reference path="./server-render.d.ts" />
 /// <reference path="./service-configuration.d.ts" />
 /// <reference path="./session.d.ts" />
+/// <reference path="./globals/session.d.ts" />
 /// <reference path="./templating.d.ts" />
+/// <reference path="./globals/templating.d.ts" />
 /// <reference path="./tiny-test.d.ts" />
+/// <reference path="./globals/tiny-test.d.ts" />
 /// <reference path="./tools.d.ts" />
+/// <reference path="./globals/tools.d.ts" />
 /// <reference path="./tracker.d.ts" />
+/// <reference path="./globals/tracker.d.ts" />
 /// <reference path="./underscore.d.ts" />
 /// <reference path="./webapp.d.ts" />
