+++ types/meteor/templating.d.ts
@@ -1,11 +1,3 @@
-
-declare var Template: TemplateStatic;
-interface TemplateStatic extends Blaze.TemplateStatic {
-    new (viewName?: string, renderFunction?: Function): Blaze.Template;
-    body: Blaze.Template;
-    [index: string]: any | Blaze.Template;
-}
-
 declare module "meteor/templating" {
     var Template: TemplateStatic;
     interface TemplateStatic extends Blaze.TemplateStatic {
