+++ types/meteor/test/meteor-tests-module-only.ts
@@ -1,0 +1,8 @@
+// This file is for tests that involve modules that don't have corresponding
+// globals.
+
+import { onPageLoad } from "meteor/server-render";
+import { ServiceConfiguration } from "meteor/service-configuration";
+import { WebApp } from "meteor/webapp";
+
+// We have no actual tests beyond the imports yet.
