+++ types/meteor/tsconfig.json
@@ -19,28 +19,48 @@
     },
     "files": [
         "accounts-base.d.ts",
+        "globals/accounts-base.d.ts",
         "blaze.d.ts",
+        "globals/blaze.d.ts",
         "browser-policy-common.d.ts",
+        "globals/browser-policy-common.d.ts",
         "check.d.ts",
+        "globals/check.d.ts",
         "ddp-rate-limiter.d.ts",
+        "globals/ddp-rate-limiter.d.ts",
         "ddp.d.ts",
+        "globals/ddp.d.ts",
         "ejson.d.ts",
+        "globals/ejson.d.ts",
         "email.d.ts",
+        "globals/email.d.ts",
         "http.d.ts",
+        "globals/http.d.ts",
         "meteor.d.ts",
+        "globals/meteor.d.ts",
         "mongo.d.ts",
+        "globals/mongo.d.ts",
         "random.d.ts",
+        "globals/random.d.ts",
         "reactive-var.d.ts",
+        "globals/reactive-var.d.ts",
         "server-render.d.ts",
         "service-configuration.d.ts",
         "session.d.ts",
+        "globals/session.d.ts",
         "templating.d.ts",
+        "globals/templating.d.ts",
         "tiny-test.d.ts",
+        "globals/tiny-test.d.ts",
         "tools.d.ts",
+        "globals/tools.d.ts",
         "tracker.d.ts",
+        "globals/tracker.d.ts",
         "underscore.d.ts",
         "webapp.d.ts",
         "index.d.ts",
-        "meteor-tests.ts"
+        "test/meteor-tests.ts",
+        "test/globals/meteor-tests.ts",
+        "test/meteor-tests-module-only.ts"
     ]
 }
